Time: 159 ms (61.61%), Space: 151 MB (81.25%) - LeetHub

+++ 3524-find-x-value-of-array-i/3524-find-x-value-of-array-i.cpp
@@ -7,7 +7,7 @@ public:
         vector<long long> count(k,0);
         count[nums[i]%k]++;
         for(int oldrem = 0; oldrem<k; oldrem++){
-            int newrem = ((long long) oldrem*nums[i]%k)%k;
+            int newrem = (oldrem*(nums[i]%k))%k;
             count[newrem]+=pre[oldrem];
         }
         pre = count;
